Tried to add the texture, but it fails.
Reading from a PNG with SDL_image doesn't seem to work, as the pixels pointer is empty. Even when I change to JPG and change the mtl to use the JPG, then change the out_color in the fragment shader to the texture, there is no more model on the screen and there are no errors in the console.